Move rustfst-tests-openfst to unit tests of rustfst in order to be able to test private interface
