--- rustfst-tests-openfst/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = 'rustfst-tests-openfst'
-version = "0.3.1-alpha.0"
-authors = ['Alexandre Caulier<alexandre.caulier@protonmail.com>']
-description = 'Library for constructing, combining, optimizing, and searching weighted finite-state transducers (FSTs).'
-documentation = 'https://docs.rs/rustfst'
-license = 'MIT/Apache-2.0'
-repository = 'https://github.com/Garvys/rustfst'
-edition = '2018'
-
-[dependencies]
-rustfst = {path = '../rustfst'}
-counter = '0.4'
-rand = '0.5'
-serde = '1.0'
-serde_derive = '1.0'
-serde_json = '1.0'
-tempfile = '3.0'
-failure = '0.1'
