+++ rustfst/Cargo.toml
@@ -34,3 +34,4 @@ rand = '0.5'
 serde = '1.0'
 serde_derive = '1.0'
 serde_json = '1.0'
+tempfile = '3.0'
