+++ rustfst/src/tests_openfst/algorithms/determinize.rs
@@ -2,15 +2,15 @@ use failure::{format_err, Fallible};
 
 use serde_derive::{Deserialize, Serialize};
 
-use rustfst::algorithms::{determinize, isomorphic, DeterminizeType};
-use rustfst::fst_properties::FstProperties;
-use rustfst::fst_traits::MutableFst;
-use rustfst::fst_traits::TextParser;
-use rustfst::semirings::Semiring;
-use rustfst::semirings::WeaklyDivisibleSemiring;
-use rustfst::semirings::WeightQuantize;
+use crate::algorithms::{determinize, isomorphic, DeterminizeType};
+use crate::fst_properties::FstProperties;
+use crate::fst_traits::MutableFst;
+use crate::fst_traits::TextParser;
+use crate::semirings::Semiring;
+use crate::semirings::WeaklyDivisibleSemiring;
+use crate::semirings::WeightQuantize;
 
-use crate::TestData;
+use crate::tests_openfst::TestData;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct DeterminizeOperationResult {
