+++ rustfst/src/tests_openfst/algorithms/minimize.rs
@@ -1,15 +1,15 @@
 use failure::{Fallible, format_err};
 use serde_derive::{Deserialize, Serialize};
 
-use rustfst::algorithms::{DeterminizeType, isomorphic, minimize};
-use rustfst::fst_properties::FstProperties;
-use rustfst::fst_traits::MutableFst;
-use rustfst::fst_traits::TextParser;
-use rustfst::semirings::Semiring;
-use rustfst::semirings::WeaklyDivisibleSemiring;
-use rustfst::semirings::WeightQuantize;
+use crate::algorithms::{DeterminizeType, isomorphic, minimize};
+use crate::fst_properties::FstProperties;
+use crate::fst_traits::MutableFst;
+use crate::fst_traits::TextParser;
+use crate::semirings::Semiring;
+use crate::semirings::WeaklyDivisibleSemiring;
+use crate::semirings::WeightQuantize;
 
-use crate::TestData;
+use crate::tests_openfst::TestData;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct MinimizeOperationResult {
