+++ rustfst/src/tests_openfst/algorithms/mod.rs
