+++ rustfst/src/tests_openfst/algorithms/topsort.rs
@@ -1,12 +1,12 @@
 use failure::Fallible;
 
-use rustfst::algorithms::top_sort;
-use rustfst::fst_properties::FstProperties;
-use rustfst::fst_traits::MutableFst;
-use rustfst::fst_traits::TextParser;
-use rustfst::semirings::Semiring;
+use crate::algorithms::top_sort;
+use crate::fst_properties::FstProperties;
+use crate::fst_traits::MutableFst;
+use crate::fst_traits::TextParser;
+use crate::semirings::Semiring;
 
-use crate::TestData;
+use crate::tests_openfst::TestData;
 
 pub fn test_topsort<F>(test_data: &TestData<F>) -> Fallible<()>
 where
