+++ rustfst/src/tests_openfst/macros.rs
